# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ClassicEntry {
  id: number;
  name: string;
  wins: number;
  games: number;
  bestTime?: number | null;
  updatedAt: string;
}

export interface ClassicSubmit {
  name: string;
  won: boolean;
  timeSeconds: number;
}

export interface InfiniteEntry {
  id: number;
  name: string;
  boards: number;
  playedAt: string;
}

export interface InfiniteSubmit {
  name: string;
  boards: number;
}