# lib/api-zod/src/generated/types/classicEntry.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ClassicEntry {
  id: number;
  name: string;
  wins: number;
  games: number;
  bestTime?: number | null;
  updatedAt: string;
}