# lib/api-zod/src/generated/types/infiniteEntry.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InfiniteEntry {
  id: number;
  name: string;
  boards: number;
  playedAt: string;
}